Moved the grid we want in GridRenderer class, and made viewports used that one
